Conflicts:
	xbmc/music/MusicThumbLoader.cpp

+++ xbmc/pvr/PVRManager.cpp
@@ -89,11 +89,26 @@ CPVRManager &CPVRManager::Get(void)
   return pvrManagerInstance;
 }
 
+bool CPVRManager::IsPVRWindowActive(void) const
+{
+  return g_windowManager.IsWindowActive(WINDOW_PVR) ||
+      g_windowManager.IsWindowActive(WINDOW_DIALOG_PVR_CHANNEL_MANAGER) ||
+      g_windowManager.IsWindowActive(WINDOW_DIALOG_PVR_OSD_CHANNELS) ||
+      g_windowManager.IsWindowActive(WINDOW_DIALOG_PVR_GROUP_MANAGER) ||
+      g_windowManager.IsWindowActive(WINDOW_DIALOG_PVR_GUIDE_INFO) ||
+      g_windowManager.IsWindowActive(WINDOW_DIALOG_PVR_OSD_CUTTER) ||
+      g_windowManager.IsWindowActive(WINDOW_DIALOG_PVR_OSD_DIRECTOR) ||
+      g_windowManager.IsWindowActive(WINDOW_DIALOG_PVR_OSD_GUIDE) ||
+      g_windowManager.IsWindowActive(WINDOW_DIALOG_PVR_GUIDE_SEARCH) ||
+      g_windowManager.IsWindowActive(WINDOW_DIALOG_PVR_RECORDING_INFO) ||
+      g_windowManager.IsWindowActive(WINDOW_DIALOG_PVR_TIMER_SETTING);
+}
+
 bool CPVRManager::InstallAddonAllowed(const std::string& strAddonId) const
 {
   return !IsStarted() ||
       !m_addons->IsInUse(strAddonId) ||
-      (!g_windowManager.IsWindowActive(WINDOW_PVR) && !IsPlaying());
+      (!IsPVRWindowActive() && !IsPlaying());
 }
 
 void CPVRManager::MarkAsOutdated(const std::string& strAddonId, const std::string& strReferer)
@@ -261,6 +276,9 @@ void CPVRManager::Stop(void)
     return;
 
   SetState(ManagerStateStopping);
+
+  if (g_windowManager.GetActiveWindow() == WINDOW_PVR)
+    g_windowManager.ActivateWindow(WINDOW_HOME);
 
   /* stop the EPG updater, since it might be using the pvr add-ons */
   g_EpgContainer.Stop();
@@ -548,22 +566,18 @@ bool CPVRManager::ContinueLastChannel(void)
   return bReturn;
 }
 
-void CPVRManager::ResetDatabase(bool bShowProgress /* = true */)
+void CPVRManager::ResetDatabase(bool bResetEPGOnly /* = false */)
 {
   CLog::Log(LOGNOTICE,"PVRManager - %s - clearing the PVR database", __FUNCTION__);
 
   g_EpgContainer.Stop();
 
-  CGUIDialogProgress* pDlgProgress = NULL;
-  if (bShowProgress)
-  {
-    pDlgProgress = (CGUIDialogProgress*)g_windowManager.GetWindow(WINDOW_DIALOG_PROGRESS);
-    pDlgProgress->SetLine(0, StringUtils::EmptyString);
-    pDlgProgress->SetLine(1, g_localizeStrings.Get(19186)); // All data in the PVR database is being erased
-    pDlgProgress->SetLine(2, StringUtils::EmptyString);
-    pDlgProgress->StartModal();
-    pDlgProgress->Progress();
-  }
+  CGUIDialogProgress* pDlgProgress = (CGUIDialogProgress*)g_windowManager.GetWindow(WINDOW_DIALOG_PROGRESS);
+  pDlgProgress->SetLine(0, StringUtils::EmptyString);
+  pDlgProgress->SetLine(1, g_localizeStrings.Get(19186)); // All data in the PVR database is being erased
+  pDlgProgress->SetLine(2, StringUtils::EmptyString);
+  pDlgProgress->StartModal();
+  pDlgProgress->Progress();
 
   if (m_addons && m_addons->IsPlaying())
   {
@@ -571,21 +585,17 @@ void CPVRManager::ResetDatabase(bool bShowProgress /* = true */)
     CApplicationMessenger::Get().MediaStop();
   }
 
-  if (bShowProgress)
-  {
-    pDlgProgress->SetPercentage(10);
-    pDlgProgress->Progress();
-  }
+  pDlgProgress->SetPercentage(10);
+  pDlgProgress->Progress();
+
+  /* reset the EPG pointers */
+  m_database->ResetEPG();
 
   /* stop the thread */
-  if (g_guiSettings.GetBool("pvrmanager.enabled"))
-    Stop();
+  Stop();
 
-  if (bShowProgress)
-  {
-    pDlgProgress->SetPercentage(20);
-    pDlgProgress->Progress();
-  }
+  pDlgProgress->SetPercentage(20);
+  pDlgProgress->Progress();
 
   if (!m_database)
     m_database = new CPVRDatabase;
@@ -593,40 +603,28 @@ void CPVRManager::ResetDatabase(bool bShowProgress /* = true */)
   if (m_database && m_database->Open())
   {
     /* clean the EPG database */
-    g_EpgContainer.Clear(true);
-    if (bShowProgress)
-    {
-      pDlgProgress->SetPercentage(30);
-      pDlgProgress->Progress();
-    }
+    g_EpgContainer.Reset();
+    pDlgProgress->SetPercentage(30);
+    pDlgProgress->Progress();
 
-    m_database->DeleteChannelGroups();
-    if (bShowProgress)
+    if (!bResetEPGOnly)
     {
+      m_database->DeleteChannelGroups();
       pDlgProgress->SetPercentage(50);
       pDlgProgress->Progress();
-    }
 
-    /* delete all channels */
-    m_database->DeleteChannels();
-    if (bShowProgress)
-    {
+      /* delete all channels */
+      m_database->DeleteChannels();
       pDlgProgress->SetPercentage(70);
       pDlgProgress->Progress();
-    }
 
-    /* delete all channel settings */
-    m_database->DeleteChannelSettings();
-    if (bShowProgress)
-    {
+      /* delete all channel settings */
+      m_database->DeleteChannelSettings();
       pDlgProgress->SetPercentage(80);
       pDlgProgress->Progress();
-    }
 
-    /* delete all client information */
-    m_database->DeleteClients();
-    if (bShowProgress)
-    {
+      /* delete all client information */
+      m_database->DeleteClients();
       pDlgProgress->SetPercentage(90);
       pDlgProgress->Progress();
     }
@@ -634,9 +632,7 @@ void CPVRManager::ResetDatabase(bool bShowProgress /* = true */)
     m_database->Close();
   }
 
-  CLog::Log(LOGNOTICE,"PVRManager - %s - PVR database cleared", __FUNCTION__);
-
-  g_EpgContainer.Start();
+  CLog::Log(LOGNOTICE,"PVRManager - %s - %s database cleared", __FUNCTION__, bResetEPGOnly ? "EPG" : "PVR and EPG");
 
   if (g_guiSettings.GetBool("pvrmanager.enabled"))
   {
@@ -646,21 +642,8 @@ void CPVRManager::ResetDatabase(bool bShowProgress /* = true */)
     Start();
   }
 
-  if (bShowProgress)
-  {
-    pDlgProgress->SetPercentage(100);
-    pDlgProgress->Close();
-  }
-}
-
-void CPVRManager::ResetEPG(void)
-{
-  CLog::Log(LOGNOTICE,"PVRManager - %s - clearing the EPG database", __FUNCTION__);
-
-  m_database->ResetEPG();
-  Stop();
-  g_EpgContainer.Reset();
-  Start();
+  pDlgProgress->SetPercentage(100);
+  pDlgProgress->Close();
 }
 
 bool CPVRManager::IsPlaying(void) const
